+++ src/rules/extensions.js
@@ -1,6 +1,8 @@
 import path from 'path'
-import resolve from '../core/resolve'
 import endsWith from 'lodash.endswith'
+
+import resolve from '../core/resolve'
+import { isBuiltIn } from '../core/importType'
 
 module.exports = function (context) {
   const configuration = context.options[0] || 'never'
@@ -24,17 +26,25 @@ module.exports = function (context) {
   function checkFileExtension(node) {
     const { source } = node
     const importPath = source.value
-    const resolvedPath = resolve(importPath, context)
-    const extension = path.extname(resolvedPath).substring(1)
 
-    if (!endsWith(importPath, extension)) {
+    // don't enforce anything on builtins
+    if (isBuiltIn(importPath)) return
+
+    const resolvedPath = resolve(importPath, context)
+
+    // get extension from resolved path, if possible.
+    // for unresolved, use source value.
+    const extension = path.extname(resolvedPath || importPath).substring(1)
+
+    if (!extension || !endsWith(importPath, extension)) {
       if (isUseOfExtensionEnforced(extension)) {
         context.report({
           node: source,
-          message: `Missing file extension "${extension}" for "${importPath}"`,
+          message:
+            `Missing file extension ${extension ? `"${extension}" ` : ''}for "${importPath}"`,
         })
       }
-    } else {
+    } else if (extension) {
       if (!isUseOfExtensionEnforced(extension) && isResolvableWithoutExtension(importPath)) {
         context.report({
           node: source,