+++ tests/src/rules/extensions.js
@@ -1,6 +1,6 @@
 import { RuleTester } from 'eslint'
-import rule from 'rules/extensions';
-import { test } from '../utils';
+import rule from 'rules/extensions'
+import { test } from '../utils'
 
 const ruleTester = new RuleTester()
 
@@ -10,28 +10,36 @@ ruleTester.run('extensions', rule, {
     test({ code: 'import dot from "./file.with.dot"' }),
     test({
       code: 'import a from "a/index.js"',
-      options: [ 'always' ]
+      options: [ 'always' ],
     }),
     test({
       code: 'import dot from "./file.with.dot.js"',
-      options: [ 'always' ]
+      options: [ 'always' ],
     }),
     test({
       code: [
         'import a from "a"',
         'import packageConfig from "./package.json"',
       ].join('\n'),
-      options: [ { json: 'always', js: 'never' } ]
+      options: [ { json: 'always', js: 'never' } ],
     }),
     test({
       code: [
         'import lib from "./bar"',
         'import component from "./bar.jsx"',
-        'import data from "./bar.json"'
+        'import data from "./bar.json"',
       ].join('\n'),
       options: [ 'never' ],
-      settings: { 'import/resolve': { 'extensions': [ '.js', '.jsx', '.json' ] } }
-    })
+      settings: { 'import/resolve': { 'extensions': [ '.js', '.jsx', '.json' ] } },
+    }),
+
+    // unresolved (#271/#295)
+    test({ code: 'import path from "path"' }),
+    test({ code: 'import path from "path"', options: [ 'never' ] }),
+    test({ code: 'import path from "path"', options: [ 'always' ] }),
+    test({ code: 'import thing from "./fake-file.js"', options: [ 'always' ] }),
+    test({ code: 'import thing from "non-package"', options: [ 'never' ] }),
+
   ],
 
   invalid: [
@@ -40,8 +48,8 @@ ruleTester.run('extensions', rule, {
       errors: [ {
         message: 'Unexpected use of file extension "js" for "a/index.js"',
         line: 1,
-        column: 15
-      } ]
+        column: 15,
+      } ],
     }),
     test({
       code: 'import a from "a"',
@@ -49,19 +57,19 @@ ruleTester.run('extensions', rule, {
       errors: [ {
         message: 'Missing file extension "js" for "a"',
         line: 1,
-        column: 15
-      } ]
+        column: 15,
+      } ],
     }),
     test({
       code: 'import dot from "./file.with.dot"',
-      options: [ "always" ],
+      options: [ 'always' ],
       errors: [
         {
           message: 'Missing file extension "js" for "./file.with.dot"',
           line: 1,
-          column: 17
-        }
-      ]
+          column: 17,
+        },
+      ],
     }),
     test({
       code: [
@@ -74,20 +82,20 @@ ruleTester.run('extensions', rule, {
         {
           message: 'Unexpected use of file extension "js" for "a/index.js"',
           line: 1,
-          column: 15
+          column: 15,
         },
         {
           message: 'Missing file extension "json" for "./package"',
           line: 2,
-          column: 27
-        }
-      ]
+          column: 27,
+        },
+      ],
     }),
     test({
       code: [
         'import lib from "./bar.js"',
         'import component from "./bar.jsx"',
-        'import data from "./bar.json"'
+        'import data from "./bar.json"',
       ].join('\n'),
       options: [ 'never' ],
       settings: { 'import/resolve': { 'extensions': [ '.js', '.jsx', '.json' ] } },
@@ -95,15 +103,15 @@ ruleTester.run('extensions', rule, {
         {
             message: 'Unexpected use of file extension "js" for "./bar.js"',
             line: 1,
-            column: 17
-        }
-      ]
+            column: 17,
+        },
+      ],
     }),
     test({
       code: [
         'import lib from "./bar.js"',
         'import component from "./bar.jsx"',
-        'import data from "./bar.json"'
+        'import data from "./bar.json"',
       ].join('\n'),
       options: [ { json: 'always', js: 'never', jsx: 'never' } ],
       settings: { 'import/resolve': { 'extensions': [ '.js', '.jsx', '.json' ] } },
@@ -111,11 +119,34 @@ ruleTester.run('extensions', rule, {
         {
             message: 'Unexpected use of file extension "js" for "./bar.js"',
             line: 1,
-            column: 17
-        }
-      ]
-    })
+            column: 17,
+        },
+      ],
+    }),
 
-  ]
+    // unresolved (#271/#295)
+    test({
+      code: 'import thing from "./fake-file.js"',
+      options: [ 'never' ],
+      errors: [
+        {
+            message: 'Unexpected use of file extension "js" for "./fake-file.js"',
+            line: 1,
+            column: 19,
+        },
+      ],
+    }),
+    test({
+      code: 'import thing from "non-package"',
+      options: [ 'always' ],
+      errors: [
+        {
+            message: 'Missing file extension for "non-package"',
+            line: 1,
+            column: 19,
+        },
+      ],
+    }),
+
+  ],
 })
-